move cancelTask from `utils` to `scheduler`
